Control::valueToClient method for R -> JS conversion
By default it does nothing, i.e. it trusts the builtin
converter. Control classes can define this method to
perform a conversion before the usual converter is called.
The DatePicker and DateRange controls use this currently,
to fix a bug when initial values of the controls are set
in the notebook.